Added limited error handling
Added limited error handling for attempting to write to the 'All Users'  profile location and receiving an access denied error. Ideally this handling could be extended to either check for admin permissions before the attempt or to correct (add permissions if fails). In this case, based on time, I am opting for neither.